+++ v3/crates/auth/hasura-authn-core/src/lib.rs
@@ -23,24 +23,81 @@ use std::{
 // Session variable and role are defined as part of OpenDD
 pub use open_dds::{
     permissions::Role,
-    session_variables::{SessionVariable, SESSION_VARIABLE_ROLE},
+    session_variables::{SessionVariableName, SessionVariableReference, SESSION_VARIABLE_ROLE},
 };
 
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, JsonSchema)]
-/// Value of a session variable
-pub struct SessionVariableValue(pub String);
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, derive_more::Display)]
+/// Value of a session variable, used to capture session variable input from parsed sources (jwt, webhook, etc)
+/// and unparsed sources (http headers)
+pub enum SessionVariableValue {
+    /// An unparsed session variable value as a string. Might be a raw string, might be a number, might be json.
+    /// How we interpret it depends on what type we're trying to coerce to from the string
+    #[display(fmt = "{_0}")]
+    Unparsed(String),
+    /// A parsed JSON session variable value. We know what the type is because we parsed it from JSON.
+    #[display(fmt = "{_0}")]
+    Parsed(serde_json::Value),
+}
 
 impl SessionVariableValue {
     pub fn new(value: &str) -> Self {
-        SessionVariableValue(value.to_string())
+        SessionVariableValue::Unparsed(value.to_string())
+    }
+
+    /// Assert that a session variable represents a string, regardless of encoding
+    pub fn as_str(&self) -> Option<&str> {
+        match self {
+            SessionVariableValue::Unparsed(s) => Some(s.as_str()),
+            SessionVariableValue::Parsed(value) => value.as_str(),
+        }
+    }
+
+    pub fn as_i64(&self) -> Option<i64> {
+        match self {
+            SessionVariableValue::Unparsed(s) => s.parse::<i64>().ok(),
+            SessionVariableValue::Parsed(value) => value.as_i64(),
+        }
+    }
+
+    pub fn as_f64(&self) -> Option<f64> {
+        match self {
+            SessionVariableValue::Unparsed(s) => s.parse::<f64>().ok(),
+            SessionVariableValue::Parsed(value) => value.as_f64(),
+        }
+    }
+
+    pub fn as_bool(&self) -> Option<bool> {
+        match self {
+            SessionVariableValue::Unparsed(s) => s.parse::<bool>().ok(),
+            SessionVariableValue::Parsed(value) => value.as_bool(),
+        }
+    }
+
+    pub fn as_value(&self) -> serde_json::Result<serde_json::Value> {
+        match self {
+            SessionVariableValue::Unparsed(s) => serde_json::from_str(s),
+            SessionVariableValue::Parsed(value) => Ok(value.clone()),
+        }
     }
 }
 
+impl From<JsonSessionVariableValue> for SessionVariableValue {
+    fn from(value: JsonSessionVariableValue) -> Self {
+        SessionVariableValue::Parsed(value.0)
+    }
+}
+
+/// JSON value of a session variable
+// This is used instead of SessionVariableValue when only JSON session variable values are accepted
+#[derive(Debug, serde::Serialize, serde::Deserialize, PartialEq, Clone, JsonSchema)]
+#[schemars(rename = "SessionVariableValue")] // Renamed to keep json schema compatibility
+pub struct JsonSessionVariableValue(pub serde_json::Value);
+
 #[derive(Clone, Debug, Eq, PartialEq, serde::Serialize)]
-pub struct SessionVariables(HashMap<SessionVariable, SessionVariableValue>);
+pub struct SessionVariables(HashMap<SessionVariableName, SessionVariableValue>);
 
 impl SessionVariables {
-    pub fn get(&self, session_variable: &SessionVariable) -> Option<&SessionVariableValue> {
+    pub fn get(&self, session_variable: &SessionVariableName) -> Option<&SessionVariableValue> {
         self.0.get(session_variable)
     }
 }
@@ -56,14 +113,14 @@ pub struct Session {
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct RoleAuthorization {
     pub role: Role,
-    pub session_variables: HashMap<SessionVariable, SessionVariableValue>,
+    pub session_variables: HashMap<SessionVariableName, SessionVariableValue>,
     pub allowed_session_variables_from_request: SessionVariableList,
 }
 
 impl RoleAuthorization {
     pub fn build_session(
         &self,
-        mut variables: HashMap<SessionVariable, SessionVariableValue>,
+        mut variables: HashMap<SessionVariableName, SessionVariableValue>,
     ) -> Session {
         let allowed_client_session_variables = match &self.allowed_session_variables_from_request {
             SessionVariableList::All => variables,
@@ -87,7 +144,7 @@ pub enum SessionVariableList {
     // like * in Select * from ...
     All,
     // An explicit list
-    Some(HashSet<SessionVariable>),
+    Some(HashSet<SessionVariableName>),
 }
 
 // Privileges of the current user
@@ -183,21 +240,21 @@ pub fn authorize_identity(
     let mut role = None;
     // traverse through the headers and collect role and session variables
     for (header_name, header_value) in headers {
-        if let Ok(session_variable) = SessionVariable::from_str(header_name.as_str()) {
-            let variable_value = match header_value.to_str() {
-                Err(e) => Err(SessionError::InvalidHeaderValue {
-                    header_name: header_name.to_string(),
-                    error: e.to_string(),
-                })?,
-                Ok(h) => SessionVariableValue::new(h),
-            };
+        let Ok(session_variable) = SessionVariableName::from_str(header_name.as_str());
+        let variable_value_str = match header_value.to_str() {
+            Err(e) => Err(SessionError::InvalidHeaderValue {
+                header_name: header_name.to_string(),
+                error: e.to_string(),
+            })?,
+            Ok(h) => h,
+        };
+        let variable_value = SessionVariableValue::Unparsed(variable_value_str.to_string());
 
-            if session_variable == SESSION_VARIABLE_ROLE {
-                role = Some(Role::new(&variable_value.0));
-            } else {
-                // TODO: Handle the duplicate case?
-                session_variables.insert(session_variable, variable_value);
-            }
+        if session_variable == SESSION_VARIABLE_ROLE {
+            role = Some(Role::new(variable_value_str));
+        } else {
+            // TODO: Handle the duplicate case?
+            session_variables.insert(session_variable, variable_value);
         }
     }
     let session = identity
@@ -220,16 +277,16 @@ mod tests {
         let mut authenticated_session_variables = HashMap::new();
 
         authenticated_session_variables.insert(
-            SessionVariable::from_str("x-hasura-user-id").unwrap(),
+            SessionVariableName::from_str("x-hasura-user-id").unwrap(),
             SessionVariableValue::new("1"),
         );
 
         client_session_variables.insert(
-            SessionVariable::from_str("x-hasura-custom").unwrap(),
+            SessionVariableName::from_str("x-hasura-custom").unwrap(),
             SessionVariableValue::new("test"),
         );
         client_session_variables.insert(
-            SessionVariable::from_str("x-hasura-custom-claim").unwrap(),
+            SessionVariableName::from_str("x-hasura-custom-claim").unwrap(),
             SessionVariableValue::new("claim-value"),
         );
         let role_authorization = RoleAuthorization {
@@ -243,7 +300,7 @@ mod tests {
         let mut expected_session_variables = client_session_variables.clone();
 
         expected_session_variables.insert(
-            SessionVariable::from_str("x-hasura-user-id").unwrap(),
+            SessionVariableName::from_str("x-hasura-user-id").unwrap(),
             SessionVariableValue::new("1"),
         );
         pa::assert_eq!(
@@ -262,23 +319,23 @@ mod tests {
         let mut authenticated_session_variables = HashMap::new();
 
         authenticated_session_variables.insert(
-            SessionVariable::from_str("x-hasura-user-id").unwrap(),
+            SessionVariableName::from_str("x-hasura-user-id").unwrap(),
             SessionVariableValue::new("1"),
         );
 
         client_session_variables.insert(
-            SessionVariable::from_str("x-hasura-custom").unwrap(),
+            SessionVariableName::from_str("x-hasura-custom").unwrap(),
             SessionVariableValue::new("test"),
         );
         client_session_variables.insert(
-            SessionVariable::from_str("x-hasura-custom-claim").unwrap(),
+            SessionVariableName::from_str("x-hasura-custom-claim").unwrap(),
             SessionVariableValue::new("claim-value"),
         );
         let mut allowed_sesion_variables_from_request = HashSet::new();
         allowed_sesion_variables_from_request
-            .insert(SessionVariable::from_str("x-hasura-custom").unwrap());
+            .insert(SessionVariableName::from_str("x-hasura-custom").unwrap());
         allowed_sesion_variables_from_request
-            .insert(SessionVariable::from_str("x-hasura-custom-author-id").unwrap());
+            .insert(SessionVariableName::from_str("x-hasura-custom-author-id").unwrap());
         let role_authorization = RoleAuthorization {
             role: Role::new("test-role"),
             session_variables: authenticated_session_variables,
@@ -291,10 +348,10 @@ mod tests {
 
         let mut expected_session_variables = client_session_variables;
         expected_session_variables
-            .remove(&SessionVariable::from_str("x-hasura-custom-claim").unwrap());
+            .remove(&SessionVariableName::from_str("x-hasura-custom-claim").unwrap());
 
         expected_session_variables.insert(
-            SessionVariable::from_str("x-hasura-user-id").unwrap(),
+            SessionVariableName::from_str("x-hasura-user-id").unwrap(),
             SessionVariableValue::new("1"),
         );
         pa::assert_eq!(
@@ -313,16 +370,16 @@ mod tests {
         let mut authenticated_session_variables = HashMap::new();
 
         authenticated_session_variables.insert(
-            SessionVariable::from_str("x-hasura-user-id").unwrap(),
+            SessionVariableName::from_str("x-hasura-user-id").unwrap(),
             SessionVariableValue::new("1"),
         );
 
         client_session_variables.insert(
-            SessionVariable::from_str("x-hasura-custom").unwrap(),
+            SessionVariableName::from_str("x-hasura-custom").unwrap(),
             SessionVariableValue::new("test"),
         );
         client_session_variables.insert(
-            SessionVariable::from_str("x-hasura-custom-claim").unwrap(),
+            SessionVariableName::from_str("x-hasura-custom-claim").unwrap(),
             SessionVariableValue::new("claim-value"),
         );
 
@@ -337,7 +394,7 @@ mod tests {
         let mut expected_session_variables = HashMap::new();
 
         expected_session_variables.insert(
-            SessionVariable::from_str("x-hasura-user-id").unwrap(),
+            SessionVariableName::from_str("x-hasura-user-id").unwrap(),
             SessionVariableValue::new("1"),
         );