+++ v3/crates/metadata-resolve/src/stages/mod.rs
@@ -27,6 +27,7 @@ mod types;
 use crate::types::warning::Warning;
 use open_dds::flags;
 pub use types::Metadata;
+pub mod scalar_type_representations;
 
 use crate::types::configuration::Configuration;
 use crate::types::error::{Error, SeparatedBy, ShouldBeAnError, WithContext};
@@ -150,6 +151,7 @@ pub fn resolve(
     let boolean_expressions::BooleanExpressionsOutput {
         boolean_expression_types,
         graphql_types,
+        issues,
     } = boolean_expressions::resolve(
         &metadata_accessor,
         boolean_expression_scalar_types,
@@ -161,6 +163,8 @@ pub fn resolve(
         &relationships,
     )
     .map_err(Error::from)?;
+
+    all_issues.extend(issues.into_iter().map(Warning::from));
 
     let order_by_expressions::OrderByExpressionsOutput {
         order_by_expressions,
@@ -271,7 +275,6 @@ pub fn resolve(
         &order_by_expressions,
         &graphql_types,
         &graphql_config,
-        configuration,
     )?;
 
     all_issues.extend(issues);
@@ -313,13 +316,17 @@ pub fn resolve(
         &commands_with_permissions,
     );
 
+    // include data connector information for each scalar type
+    let scalar_types_with_representations =
+        scalar_type_representations::resolve(&data_connector_scalars, &scalar_types);
+
     let plugin_configs = plugins::resolve(&metadata_accessor);
 
     let all_warnings = warnings_as_errors_by_compatibility(&metadata_accessor.flags, all_issues)?;
 
     Ok((
         Metadata {
-            scalar_types,
+            scalar_types: scalar_types_with_representations,
             object_types: object_types_with_relationships,
             models,
             commands,