+++ v3/crates/open-dds/src/flags.rs
@@ -32,6 +32,15 @@ pub struct Flags {
 
     #[opendd(default, rename = "allow_partial_supergraph")]
     pub allow_partial_supergraph: bool,
+
+    #[opendd(default, rename = "json_session_variables")]
+    pub json_session_variables: bool,
+
+    #[opendd(
+        default,
+        rename = "disallow_array_field_compared_with_scalar_boolean_type"
+    )]
+    pub disallow_array_field_compared_with_scalar_boolean_type: bool,
 }
 
 impl Flags {
@@ -46,6 +55,8 @@ impl Flags {
             propagate_boolean_expression_deprecation_status: false,
             require_unique_command_graphql_names: false,
             allow_partial_supergraph: false,
+            json_session_variables: false,
+            disallow_array_field_compared_with_scalar_boolean_type: false,
         }
     }